feat(api): translate errors to http code errors
