+++ api/src/evaluation/domain/errors.js
@@ -11,21 +11,4 @@ class StageWithLinkedCampaignError extends DomainError {
   }
 }
 
-class AutonomousCourseRequiresATargetProfileWithSimplifiedAccessError extends DomainError {
-  constructor() {
-    super('Autonomous course requires a target profile with simplified access.');
-  }
-}
-
-class TargetProfileRequiresToBeLinkedToAutonomousCourseOrganization extends DomainError {
-  constructor() {
-    super('Target profile requires to be linked to autonomous course organization.');
-  }
-}
-
-export {
-  AutonomousCourseRequiresATargetProfileWithSimplifiedAccessError,
-  DomainError,
-  StageWithLinkedCampaignError,
-  TargetProfileRequiresToBeLinkedToAutonomousCourseOrganization,
-};
+export { DomainError, StageWithLinkedCampaignError };
